requirements.txt
Flask>=3.0.0
Flask-WTF>=1.2.0
beautifulsoup4>=4.12.2
requests>=2.31.0
python-dotenv>=1.0.0
rich>=13.5.2
waitress>=2.1.2